# buffruneio: make 4X faster by using slice instead of container/list