Hi,

I noticed that a lot of Go packages are using this package for buffered rune input, and it seemed inefficient to be using container/list instead of a slice, so I changed the implementation to use a slice. I found and fixed a bug in Forget in the process, and I revised the doc comments a bit.

Thanks for writing a package that so many people want to use. :-)

```
name             old time/op    new time/op    delta
Read16K-8           886µs ± 1%     200µs ± 0%   -77.47%  (p=0.000 n=9+8)
ReadForget16K-8     944µs ± 4%     191µs ± 4%   -79.82%  (p=0.000 n=10+9)
ReadRewind16K-8     987µs ± 1%     278µs ± 1%   -71.87%  (p=0.000 n=9+10)

name             old speed      new speed      delta
Read16K-8        18.5MB/s ± 1%  82.0MB/s ± 1%  +343.56%  (p=0.000 n=9+9)
ReadForget16K-8  17.4MB/s ± 4%  86.0MB/s ± 4%  +395.53%  (p=0.000 n=10+9)
ReadRewind16K-8  16.6MB/s ± 1%  59.0MB/s ± 1%  +255.49%  (p=0.000 n=9+10)

name             old alloc/op   new alloc/op   delta
Read16K-8           791kB ± 0%     208kB ± 0%   -73.73%  (p=0.002 n=8+10)
ReadForget16K-8     791kB ± 0%       5kB ± 0%   -99.34%  (p=0.000 n=10+10)
ReadRewind16K-8     791kB ± 0%     208kB ± 0%   -73.74%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
Read16K-8           24.6k ± 0%      0.0k ± 0%   -99.94%  (p=0.000 n=10+10)
ReadForget16K-8     24.6k ± 0%      0.0k ± 0%   -99.99%  (p=0.000 n=10+10)
ReadRewind16K-8     24.6k ± 0%      0.0k ± 0%   -99.94%  (p=0.000 n=10+10)
```

https://perf.golang.org/search?q=upload:20180119.1