@rsc 2018-01-19
I see the failures with very old versions of Go due to use of strings.Reader.Reset.

I'm a little skeptical that you need this to run in Go 1.3 anymore, but Go 1.6 is still a concern due to App Engine, and strings.Reader.Reset was added in Go 1.7, so I rewrote the benchmark to avoid it.