@pelletier 2018-01-19
Thank you very much for this very thorough patch!

I had no idea anyone was using this package. As far as I knew, only [go-toml](https://github.com/pelletier/go-toml) used to until I replaced it with a [similar slice implementation](https://github.com/pelletier/go-toml/commit/69d355db5304c0f7f809a2edc054553e7142f016). Curious how you found out other projects relied on it. :)

Good catch on the `Forget` bug!

There is no reason tests should run on Go 1.3 — merely an artifact of the project not being actively worked on. I usually only support the last two major Go versions, but never realized App Engine is still stuck on 1.6. I will update the config to run on 1.6 through 1.9.