@rsc 2018-01-19
https://godoc.org/github.com/pelletier/go-buffruneio?importers shows packages importing this one. I found it via kevinburke/ssh_config, which is used by src-d/go-git.